chore(cli): Add CLI files
