+++ src/main.rs
@@ -1,6 +1,7 @@
 mod backend;
 mod cli;
 mod reg;
+mod tool;
 fn main() {
     println!("Hello, world!");
 }
