+++ src/reg/tool_reg.rs
@@ -1,7 +1,7 @@
 use anyhow::Result;
 use once_cell::sync::Lazy;
 
-static ToolReg: Cache<T> = Lazy::new(|| ToolRegAPI::get_or_init()?);
+static ToolReg: Cache<> = Lazy::new(|| ToolRegAPI::get_or_init()?);
 
 #[derive(Default, Debug)]
 pub struct ToolRegConfig {}
@@ -10,6 +10,4 @@ pub struct ToolRegAPI;
 
 impl ToolRegAPI {
     type ctx = Cache;
-
-    pub fn get_or_init() -> Result {}
 }
